solver: add sqlcachestorage as an alternative to the bolt db
This creates an sql cache storage that stores the cache key index in an
sqlite3 database instead of in bolt. At the moment, this is just
functional in the bare minimum as a way to test the feasibility of this
method. There are ways to configure the sqlite database for more
efficient transactions.

This can be used as an alternative storage for bolt db and is
potentially usable as a distributed cache key storage for another
implementation in the future.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>