bklog: always enable trace id if it exists
Previous version of the detect package would only enable trace ids in
the build log if tracing was in use. For practical purposes, there was
always a tracer in use (either the recorder which was always enabled or
one of the detected tracers).

When this package was refactored, the call to enable the trace id was
accidentally removed.

This removes the option to enable or disable the trace id from the log
and instead just always enables it since this was the behavior
previously.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>