readme: add r2d4/llb frontend and dacc project
r2d4/llb is a frontend that exposes the Solve rpc and the image config via JSON. The idea is that I found it difficult to generate the full gRPC stubs for languages other than Go. Additionally, some runtimes (like connect/grpc node/web) can't listen on a unix socket. Instead, you only have to compile ops.proto and build the graph.

The dacc project is a TypeScript project that uses the r2d4/llb frontend. It is similiar to llb.State, but in TypeScript and calls the frontend to solve the graph.

Signed-off-by: Matt Rickard <git@matt-rickard.com>