solver: make digest computation for ops deterministic again
The solver relied on protobuf marshaling to compute a digest. There's no
guarantee of the byte order, especially for maps, with marshaling. As an
implementation detail, gogo would sort the keys for a map and
`Deterministic: true` causes the standard marshaler to sort the keys,
but this isn't a reliable method.

In the future, we likely want to either remove the maps from the
protobuf messages or not rely on protobuf serialization for this
function.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>