dockerfile: set error location for ONBUILD errors
Set error location for ONBUILD errors to the command
that pulled in the image with ONBUILD definitions.

Currently location was incorrectly set to the first
line as ONBUILD was parsed as a separate source code.

Location is handled both at parse time (for invalid
ONBUILD definition) and run time (when command set
via ONBUILD errors).

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>