grpcclient: return proper nil reference from grpcclient
The grpcclient would take an empty reference and convert it into a
`(*reference)(nil)` and then store that in a `client.Reference`. This
caused the `nil` check in `convertRef` to return false because it wasn't
`nil` but then it panicked because it was `nil`.

`newReference` has been minorly refactored to not return an error. The
method is not exported, the error value is not used, and it obscured the
functionality which made it harder to use correctly.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>