detect: sever semconv relationship to otel sdk
Copies over the telemetry sdk detection to utilize our own version of
semconv rather than mix the versions between different packages. This
will keep a consistent schema url instead of constantly chasing
whichever one the otel sdk is using.

The only thing left from the otel sdk is `WithFromEnv` which is
schemaless and won't conflict for that reason so we can continue to use
it.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>