detect: use newer semconv for resource and add unit test
We upgraded a dependency that upgraded otel and also changed the semconv
schema for us to a newer version. We forgot to upgrade our own semconv.

In the future, we may want to find a way to have a detector that doesn't
require us to manually specify a semconv so we don't have to remember
this. For now, I've added a test to catch when this happens so it
doesn't happen again.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>