solver: return errors hit during cache export
There were two lines that checked if err != nil and return nil error.
This seems likely to be a typo, especially since there is support for
ignoring errors during cache export but on a completely different level
of abstraction (in llbsolver).

There were bugs causing errors during cache export (fixed in subsequent
commits) which ended up getting silently dropped and causing cache
exports to be missing mysteriously. Now errors are returned and only
ignored if cache export errors are configured to be ignored.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>