solver: allow LoadCache to set CacheOpts
For lazy remote cache cases, we figure out the descriptor handlers to
use during loading of cache rather than during a CacheMap operation.

In order to make those descHandlers available as CacheOpts we need to
plumb them through to the shared op and allow withAncestorCacheOpts to
check those in addition to the CacheOpts from CacheMaps.

This allows loading of lazy refs during cache export when there are refs
resolved with cache imports.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>