cache: add tracing to prune
Adds tracing to the cache manager prune operation. The attribute will
show both explicit prunes from a user and background prunes initiated
after solve requests are processed.

The prune code has also been refactored a bit to remove a recursive call
to prune. It has now been unrolled and the functions have been renamed
so `prune` stands for the entire prune process and `pruneOnce` is a
single cycle. Previously, `prune` was a single cycle and `pruneOnce`
would run the entire prune cycle with a set of options.

This also disabled the delve component of the `hack/compose` script. It
appears to be broken at the moment.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>