dockerfile: don't allow non-Dockerfile syntax for directives
ParseDirectives code was changed when "check" directive was added and
copied over logic from DetectSyntax function. This does not look correct
as the only allowed formatting for Dockerfile directives is with a
Dockerfile comment. "#syntax" is a special case because we want to
allow frontend forwarding capability also in non-Dockerfile sources that
use different style of comments or using config files with JSON as
frontend entrypoints.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>