fix: util/network: add option not to re-use namespaces from pool
There is an issue with `HnsEndpoints` on WS2019, after first time
use, they can't be re-attached. See #5668. HCS CreateComputeSystem
fails with:

```
The requested operation for attach namespace failed.
```

As a work-around, have an option for PoolSize=-1 to turn off the
namespace pooling functionality and have each namespace only be used
once and discarded.

To enable this, you run buildkit with the additional flag:
`--containerd-cni-pool-size=-1`.

Fixes #5668

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>