fix: docs: windows: enhance the CNI setup instructions
The previous instructions had a slight mistake where the CNI
version was similar to the CNI plugin Version. At that time,
that was true, we had 0.3.0. However, the latest plugin
version is 0.3.1 and supports upto CNI v 1.0.0.

Fix the installation scripts to have `$cniVersion` and
`$cniPluginVersion` as separate variables.

Also add script to set up HNS Network on WS2019 when
it is not already available by default, like in the
rest of the SKUs WS2022+, after enabling Hyper-V.

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>