bboltcachestorage: delete links outside of cursor
The `emptyBranchWithParents` method could accidentally leave link
entries that shouldn't exist. When finding these links, deleting during
the iteration can sometimes cause the cursor to jump entries that should
be deleted.

This changes the code path to delete the links outside of the iteration
to avoid this.

This is caused by a long-standing bug in bolt that can't be fixed
easily. See https://github.com/etcd-io/bbolt/pull/611 for details.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>