buildctl: set requestid frontend attrs before build function
Setting the frontend attributes for requestid in the build callback is
too late for it to be recognized by the controller. The frontend
attributes have already been sent by the build function to the
controller and so changing them leads to an inconsistency in the request
that can make the verifier fail.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>