verifier: do not run invalid platforms check when there are no results
When `--call=lint` or `--call=outline` were used, the intention wasn't
to run this check because the result doesn't have a platform. When those
checks are run, there is no exporter platforms key.

We would check `Request` from the request operations, but this wasn't a
reliable way to determine whether the result was from a lint because it
is possible for subrequests to have different frontend attributes than
the primary request and the primary request would sometimes not have the
`requestid` attribute set.

Instead of detecting a subrequest with that method, the check has been
updated to exit if there are no references and if there is no exporter
key.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>