cache: change the default cache mode to max
Usually whoever sets buildkit's cache options wants to use the cache as much as possible. The 1% of users that want to use the cache but not too much should explicitly set `mode=min`.

Many users think they set up the cache properly because it works but don't it could be so much better and end up having a worse experience.

The mode setting is not obvious to all users, let's make the default sane.

Signed-off-by: Yarden Shoham <git@yardenshoham.com>