uploadprovider: remove recursion from uploadprovider write code
Removes the recursive invocation of `Write` from the `Write`
implementation that chunked the bytes message writes.

This is better suited for a non-recursive algorithm. The recursive
version will create a new stackframe for each chunk which can be a
problem from a performance and stability perspective.

The code is now changed to a simple for loop that writes each chunk.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>