exporter: workaround containerd client issue with grpc size limits
The grpc client for containerd won't properly break up the writes for a
file when used in combination with `bytes.NewReader` and
`content.WriteBlob`.

To workaround the issue, this wraps the container image exporter for
containerd with a bytes reader that will chunk the `Write` calls instead
of relying on the client to do the chunking.

See https://github.com/containerd/containerd/issues/11440 for the
upstream containerd issue.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>