control: forward traces in a non-blocking goroutine
Modify the trace forwarder in the controller to process the traces in
another goroutine to prevent the original client from blocking when the
downstream exporter isn't available.

If buildkit was configured to export traces, it would proxy the traces
from a client. This export was blocking and caused the upstream client
to stall until buildkit had successfully sent the trace to its external
collector.

This changes the forwarding to be non-blocking from the client's
perspective.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>