fix: wcow: simplify CNI setup for windows
To enable CNI networking for WCOW, you needed to provide
long paths when running the daemon, e.g.
```
buildkitd `
    --containerd-cni-config-path="C:\Program Files\containerd\cni\conf\0-containerd-nat.conf" `
    --containerd-cni-binary-dir="C:\Program Files\containerd\cni\bin"
```

Fix this so that the default paths are set in appdefaults,
instead of the one that had be set previously.
Now you can just run: `buildkitd`.

Also, refactor the script for setting up CNI binaries
and configs, to make it easy to refresh the configs
without having to download the binaries afresh.

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>