wcow: leave default PATH env to be set by the OS
The PATH is very critical during container runs on
Windows. Windows stores the PATH details in its
registry hive, while in unix, this is often
stored in the image's config. See further details
at #5445

Setting a default path like we do on Linux
(which is mostly not used since the PATH is
already set in the configs), works against
the expected build experience, especially when
it comes to installers and commands like `setx`.

Therefore, we skip setting the default PATH on
Windows, and leave it for the OS to load it
from its registry hive.

This also further supports backward compatibilitiy
with the current experience with docker classic
builder.

Users wishing to explicitly store this in the
configs can opt-in by using the ENV PATH= ..
in the Dockerfile, etc.

See also the same practices on Docker Engine
and Containerd:
- https://github.com/moby/moby/blob/da3b31fb2c4e83e77cf928e166982d03ba68a4e7/oci/defaults.go#L24-L33
- https://github.com/containerd/containerd/blob/041743e8afd2bb2189aa904f04bf87b9073892d3/pkg/oci/spec_opts_windows.go#L66-L69

closes #5445

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>