gha cache: reload expired cache URLs
In some conditions there can be a pause between initial load of the
entry (that already contains signed download URL even though we don't
need it yet) and by the time cache is actually downloaded in the end
of the build. One such case seems to be accessing SBOM scanner layers
after the main build is completed. Signed URL is only valid for 10min
and if more time has passed then old URL may receive 403 error.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>