fix: display error details on unexpected response status code errors
When receiving a remote error due to an unexpected status code, read the
body and try to decode either a docker.Error or a simpler {"details"...}
payload. If those succeed, decode and format them appropriately. Otherwise,
fallback to displaying the raw response body which might contain valuable
information to indicate why authentication failed.

Signed-off-by: Alberto Garcia Hierro <damaso.hierro@docker.com>