[cache] Fix descriptor duplication in getAvailableBlobs
The walkBlob function was called with a function having a side effect: appending a descriptor to a list. However, the function could be called multiple times for the same descriptor (once in `walkBlob` and once in `walkBlobVariantsOnly` in that case) leading to duplicate entries in the final list.
Instead of the list, we can use a map to store the descriptors. With the digest as key, it will ensure we don't store duplicates.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>