[cache] Fix descriptor duplication in getAvailableBlobs
The walkBlob function was called with a function having a side effect: appending a descriptor to a list. However, the function could be called multiple times for the same descriptor (once in `walkBlob` and once in `walkBlobVariantsOnly` in that case) leading to duplicate entries in the final list.
Instead, we will check each time we want to add an element, that it's not already in the list using slice.ContainsFunc.
This should be fine from a performance perspective as the list is small (at most 2 compressions in the vast majority of cases).

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>