dap: add ability to keep ops unique when marshaling
Marshaling with the `WithIdentities()` constraint will ensure operations
are given unique identities to ensure they marshal with different
values. This prevents the same operation from being merged with other
identical operations.

This is useful for the debug adapter protocol which wants to know the
original locations of operations when they are in unique places.

This behavior can be enabled for dockerfiles by setting
`BUILDKIT_WITH_IDENTITIES`.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>