docs: use latest "labs" for --parents and --exclude
The `--parents` and `--exclude` options have been part of the labs
channel since 1.7, but have not yet been promoted to stable.

Using the 1.7-labs version in these examples means that users would
be downgrading their Dockerfile syntax from 1.17 (soon 1.18.0) to
a much older syntax.

Let's update the notes to use the latest labs syntax so that users
are not surprised by other features no longer being supported when
using the labs syntax.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>