contenthash: do not follow nested symlinks when computing checksum
Nested symlinks were not intended to be followed after the first level
of wildcard expansion. When include/exclude patterns were used or a
wildcard was present higher up in the tree, we would erroneously follow
ALL symlinks to compute the hash.

When a symlink was broken, this would result in an error. We weren't
supposed to be following the symlinks at all.

This fixes things by changing `includedPaths` to also return whether a
path should follow links. If a parent directory exists for the path,
then the symlink is not followed.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>