frontend: add required include paths to LLB and use it with --parents
This adds an additional `RequiredIncludePaths` that is primary intended
for use with `COPY --parents`. This version of include paths will verify
that the pattern matched at least one file and checksum will return a
not found error if one of the patterns did not match any files.

As part of this, it is allowed to match the path and then exclude the
same path. As long as some path would have matched without the
exclusion, then the directory is considered to exist.

This fixes an issue with `COPY --parents` where a non existant directory
that was intended to be copied would be ignored.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>