contenthash: change how checksum is calculated with wildcards and patterns
This changes the way the checksum is calculated for content that has
wildcards or include/exclude paths. The previous method would result in
a different checksum with the same file contents but different patterns
due to a different method of hashing the data.

This normalizes the way the data is hashed to use similar methods. The
wildcard path now strips the prefix from the path instead of using the
base path. It no longer adds an additional zero byte between entries but
instead uses the zero byte from the directory path key.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>